requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.0 ; python_version >= '3.7'
asn1crypto==1.4.0
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.10.0
bleach==5.0.1 ; python_version >= '3.7'
blinker==1.5 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
boto3==1.21.4
botocore==1.24.4
bs4==0.0.1
cachetools==5.2.0 ; python_version ~= '3.7'
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3 ; python_version >= '3.7'
commonmark==0.9.1
cryptography==36.0.1
decorator==5.1.1 ; python_version >= '3.5'
emoji==1.6.3
entrypoints==0.4 ; python_version >= '3.6'
frozenlist==1.3.0
gitdb==4.0.9
gitpython==3.1.27
greenlet==1.1.2
idna==3.3
importlib-metadata==5.0.0 ; python_version >= '3.7'
jinja2==3.1.2 ; python_version >= '3.7'
jmespath==0.10.0
jsonschema==4.17.0 ; python_version >= '3.7'
lxml==4.8.0
markupsafe==2.1.1 ; python_version >= '3.7'
multidict==6.0.2
numpy==1.22.2
oscrypto==1.2.1
packaging==21.3 ; python_version >= '3.6'
pandas==1.4.1
pillow==9.3.0 ; python_version >= '3.7'
protobuf==3.20.3 ; python_version >= '3.7'
pyarrow==10.0.0 ; python_version >= '3.7'
pycparser==2.21
pycryptodome==3.14.1
pycryptodomex==3.14.1
pydeck==0.8.0 ; python_version >= '3.7'
pygments==2.13.0 ; python_version >= '3.6'
pyjwt==2.3.0
pympler==1.0.1 ; python_version >= '3.6'
pyopenssl==21.0.0
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
pyrsistent==0.19.2 ; python_version >= '3.7'
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
requests==2.27.1
requests-toolbelt==0.10.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rich==12.6.0 ; python_full_version >= '3.6.3' and python_full_version < '4.0.0'
s3transfer==0.5.1
semver==2.13.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
setuptools==65.5.1 ; python_version >= '3.7'
six==1.16.0
slack-sdk==3.15.1
smmap==5.0.0
snowflake-connector-python==2.7.4
soupsieve==2.3.1
sqlalchemy==1.4.31
streamlit==1.14.1
streamlit-aggrid==0.3.3
tableau-api-lib==0.1.50
tableauhyperapi==0.0.15888
teradatasql==17.10.0.6
toml==0.10.2 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
toolz==0.12.0 ; python_version >= '3.5'
tornado==6.2 ; python_version >= '3.7'
typeguard==2.13.3 ; python_full_version >= '3.5.3'
typing-extensions==4.4.0 ; python_version >= '3.7'
tzdata==2022.6 ; python_version >= '3.6'
tzlocal==4.2 ; python_version >= '3.6'
urllib3==1.26.8
validators==0.20.0 ; python_version >= '3.4'
webencodings==0.5.1
yarl==1.7.2
zipp==3.10.0 ; python_version >= '3.7'
